Paths: Move paths to utils/
