+++ pisek/cms/testcase.py
@@ -18,7 +18,7 @@ from itertools import chain
 
 from pisek.env.env import Env
 from pisek.env.task_config import TaskType
-from pisek.paths import TaskPath
+from pisek.utils.paths import TaskPath
 
 
 def create_testcase(
