+++ pisek/jobs/parts/compile.py
@@ -21,7 +21,7 @@ import subprocess
 import sys
 
 from pisek.env.env import Env
-from pisek.paths import TaskPath
+from pisek.utils.paths import TaskPath
 from pisek.jobs.jobs import State, PipelineItemFailure
 from pisek.jobs.parts.program import ProgramsJob
