+++ pisek/jobs/parts/data.py
@@ -19,7 +19,7 @@ from typing import Any, Callable
 
 from pisek.jobs.jobs import Job, PipelineItemFailure
 from pisek.env.env import Env
-from pisek.paths import TaskPath
+from pisek.utils.paths import TaskPath
 from pisek.jobs.parts.task_job import (
     TaskJob,
     TaskJobManager,
