+++ pisek/jobs/parts/generator.py
@@ -21,7 +21,7 @@ import shutil
 from typing import Any
 
 from pisek.env.env import Env
-from pisek.paths import TaskPath, GENERATED_SUBDIR
+from pisek.utils.paths import TaskPath, GENERATED_SUBDIR
 from pisek.env.task_config import ProgramType
 from pisek.jobs.jobs import Job, PipelineItemFailure
 from pisek.jobs.parts.task_job import TaskJob, TaskJobManager
