+++ pisek/jobs/parts/judge.py
@@ -22,7 +22,7 @@ from typing import Optional, Union, Callable
 import subprocess
 
 from pisek.env.env import Env
-from pisek.paths import TaskPath
+from pisek.utils.paths import TaskPath
 from pisek.env.task_config import ProgramType, JudgeType
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
 from pisek.utils.text import tab
