+++ pisek/utils/paths.py
@@ -16,9 +16,11 @@
 
 from dataclasses import dataclass
 import os
+from typing import TYPE_CHECKING
 import yaml
 
-from pisek.env.env import Env
+if TYPE_CHECKING:
+    from pisek.env.env import Env
 
 BUILD_DIR = "build/"
 
@@ -70,82 +72,82 @@ class TaskPath:
         return TaskPath(os.path.relpath(os.path.join(*path), "."))
 
     @staticmethod
-    def static_path(env: Env, *path: str) -> "TaskPath":
+    def static_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath(env.config.static_subdir, *path)
 
     @staticmethod
-    def solution_path(env: Env, *path: str) -> "TaskPath":
+    def solution_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath(env.config.solutions_subdir, *path)
 
     @staticmethod
-    def executable_path(env: Env, *path: str) -> "TaskPath":
+    def executable_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath(BUILD_DIR, *path)
 
     @staticmethod
-    def executable_file(env: Env, program: str) -> "TaskPath":
+    def executable_file(env: "Env", program: str) -> "TaskPath":
         program = os.path.splitext(os.path.basename(program))[0]
         return TaskPath.executable_path(env, program)
 
     @staticmethod
-    def data_path(env: Env, *path: str) -> "TaskPath":
+    def data_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath(env.config.data_subdir, *path)
 
     @staticmethod
-    def generated_path(env: Env, *path: str) -> "TaskPath":
+    def generated_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, GENERATED_SUBDIR, *path)
 
     @staticmethod
-    def generated_input_file(env: Env, subtask: int, seed: int) -> "TaskPath":
+    def generated_input_file(env: "Env", subtask: int, seed: int) -> "TaskPath":
         return TaskPath.generated_path(env, f"{subtask:02}_{seed:x}.in")
 
     @staticmethod
-    def input_path(env: Env, *path: str) -> "TaskPath":
+    def input_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, INPUTS_SUBDIR, *path)
 
     @staticmethod
-    def invalid_path(env: Env, *path: str) -> "TaskPath":
+    def invalid_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, INVALID_OUTPUTS_SUBDIR, *path)
 
     @staticmethod
-    def invalid_file(env: Env, name: str, seed: int) -> "TaskPath":
+    def invalid_file(env: "Env", name: str, seed: int) -> "TaskPath":
         name = os.path.splitext(name)[0]
         return TaskPath.invalid_path(env, f"{name}.{seed:x}.invalid")
 
     @staticmethod
-    def output_path(env: Env, *path: str) -> "TaskPath":
+    def output_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, OUTPUTS_SUBDIR, *path)
 
     @staticmethod
-    def output_file(env: Env, input_name: str, solution: str) -> "TaskPath":
+    def output_file(env: "Env", input_name: str, solution: str) -> "TaskPath":
         input_name = os.path.splitext(os.path.basename(input_name))[0]
         solution = os.path.basename(solution)
         return TaskPath.output_path(env, f"{input_name}.{solution}.out")
 
     @staticmethod
-    def output_static_file(env: Env, name: str) -> "TaskPath":
+    def output_static_file(env: "Env", name: str) -> "TaskPath":
         name = os.path.splitext(name)[0]
         return TaskPath.output_path(env, f"{name}.out")
 
     @staticmethod
-    def sanitized_path(env: Env, *path: str) -> "TaskPath":
+    def sanitized_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, SANITIZED_SUBDIR, *path)
 
     @staticmethod
-    def sanitized_file(env: Env, name: str) -> "TaskPath":
+    def sanitized_file(env: "Env", name: str) -> "TaskPath":
         name = os.path.basename(name)
         return TaskPath.sanitized_path(env, f"{name}.clean")
 
     @staticmethod
-    def log_path(env: Env, *path: str) -> "TaskPath":
+    def log_path(env: "Env", *path: str) -> "TaskPath":
         return TaskPath.data_path(env, LOG_SUBDIR, *path)
 
     @staticmethod
-    def points_file(env: Env, name: str) -> "TaskPath":
+    def points_file(env: "Env", name: str) -> "TaskPath":
         name = os.path.splitext(name)[0]
         return TaskPath.log_path(env, f"{name}.points")
 
     @staticmethod
-    def log_file(env: Env, name: str, program: str) -> "TaskPath":
+    def log_file(env: "Env", name: str, program: str) -> "TaskPath":
         name = os.path.splitext(os.path.basename(name))[0]
         program = os.path.basename(program)
         return TaskPath.log_path(env, f"{name}.{program}.log")
